+++ pkg/local_object_storage/metabase/delete.go
@@ -161,13 +161,8 @@ func (db *DB) delete(tx *bbolt.Tx, addr oid.Address, currEpoch uint64) (bool, bo
 	}
 
 	// if object is an only link to a parent, then remove parent
-	if parent := obj.Parent(); parent != nil {
-		if id := parent.GetID(); id.IsZero() {
-			// unfinished header from the first part
-			return false, false, 0, nil
-		}
-
-		err = db.deleteObject(tx, obj, true)
+	if parent := obj.Parent(); parent != nil && !parent.GetID().IsZero() {
+		err = db.deleteObject(tx, parent, true)
 		if err != nil {
 			return false, false, 0, fmt.Errorf("could not remove parent object: %w", err)
 		}